WIP: Removed most usages of layer index
With the introduction of GroupLayer, all code that was solely using
a layer index to refer to some layer needs to be updated. In most cases,
the layer index was replaced by a layer pointer. However, a large amount
of code will still need to be adjusted to work on a layer hierarchy
rather than a list of layers.

This change also makes the LayerModel largely aware of GroupLayers, so
it is now a tree model.